+++ tests/pure/es.string.substr.js
@@ -6,6 +6,8 @@ QUnit.test('String#substr', assert => {
 
   assert.same(substr('12345', 1, 3), '234');
 
+  assert.same(substr('ab', -1), 'b');
+
   /* eslint-disable es/no-symbol -- safe */
   if (typeof Symbol === 'function') {
     assert.throws(() => substr(Symbol(), 1, 3), 'throws on symbol context');