+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Changelog
 ##### Unreleased
+- Fixed `String.prototype.substr` feature detection and compat data
 - Families of `.at` and `.findLast` methods marked as supported in Safari TP
 
 ##### 3.18.0 - 2021.09.20